requirements.txt
piexif==1.1.3
Pillow==11.0.0
PyQt5==5.15.11
PyQt5_sip==12.15.0
